Get all components working as modules.
